Consolidate AsFd instances for stdio types into `library/std/src/os/fd/owned.rs`
